[Fix] QuestionService 기능 수정 (#18)
문의사항 삭제 기능 - userIdx 일치하는 사용자만 문의사항 삭제 가능하도록 수정

+++ backend/src/main/java/com/example/Surisuri_Masuri/question/service/QuestionService.java
@@ -95,13 +95,15 @@ public class QuestionService {
     }
 
 
-    public BaseResponse delete(Integer idx) {
+    public BaseResponse delete(User user, Integer idx) {
         Optional<Question> result = questionRepository.findById(idx);
 
         if(result.isPresent()) {
-            Question question = result.get();
-            questionRepository.delete(question);
-            return BaseResponse.successResponse("문의사항 삭제 성공",null);
+            if (result.get().getUser().getIdx().equals(user.getIdx())) {
+                Question question = result.get();
+                questionRepository.delete(question);
+                return BaseResponse.successResponse("문의사항 삭제 성공", null);
+            }
         }
         return BaseResponse.successResponse("삭제할 문의사항이 존재하지 않습니다", null);
